+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ----------------
 
 - Add CSS classes to listing summary fields.
+- Fix wrong title for agent contact portlet add and edit forms.
+- Update german translations.
 
 
 1.6.1 (2016-02-18)